feat: Change making pdf file test location
