+++ src/page/Login/index.tsx
@@ -1,0 +1,6 @@
+import React, { useState, useEffect } from "react";
+
+function Login() {
+  return <div>Login page</div>;
+}
+export default Login;
