--- src/page/Main/index.tsx
@@ -1,12 +1,0 @@
-import React, { useState, useEffect } from "react";
-import { Header } from "../../components/basis";
-function Main() {
-  return (
-    <div>
-      <Header />
-      Main page
-    </div>
-  );
-}
-
-export default Main;
